refactor: extract process/port/browser helpers to myco-shared
Pulls the duplicated platform helpers between packages/myco's eviction
module and packages/myco-hub into a single internal source-only package
at packages/myco-shared. No new published package, no runtime dep change
— TS path aliases (`@myco-shared/*`) wire the source into both consumers
and their bundlers (bun for myco, tsup for myco-hub) inline the result.

Shared modules:
- process.ts — isProcessAlive, readProcessCwd, readProcessCommandLine,
  findVaultFromCwd/CommandLine/ForProcess (Linux + Darwin + Windows).
- port.ts — findPidsListeningInRange, findPidsListeningOn, lsof/ss/
  netstat parsers, Get-NetTCPConnection JSON parser, PORT_RANGE
  constants.
- terminate.ts — terminateProcess (SIGTERM→SIGKILL ladder),
  cleanStaleDaemonJson.
- open-browser.ts — single spawn-based implementation (replaces myco's
  shell-exec variant).
- json.ts — tryParseJson + readJsonFile.

Consumer changes:
- packages/myco/src/daemon/eviction.ts shrinks 379 → 209 lines; keeps the
  vault-aware public surface (evictDaemonsForVault,
  findDaemonTargetsForVault, isMycoDaemonForVault) and re-exports the
  shared helpers so tests/CLI callers see no API churn.
- packages/myco-hub/src/process.ts becomes a re-export façade.
- packages/myco-hub/src/daemon.ts:stopProject collapses onto
  terminateProcess + cleanStaleDaemonJson, removing the local
  SIGTERM/SIGKILL/poll dance and an unused waitForExit helper.
- packages/myco-hub/src/discovery.ts uses PORT_RANGE_START/END from
  shared and switches readDaemonJson onto readJsonFile.
- packages/myco/src/utils/json.ts re-exports from shared so existing
  `@myco/utils/json` import path keeps working.

50 hub + eviction tests pass; full tsc clean.
