+++ packages/myco-hub/src/open-browser.ts
@@ -1,12 +1,1 @@
-import { spawn } from 'node:child_process';
-
-export function openBrowser(url: string): void {
-  const command = process.platform === 'darwin'
-    ? 'open'
-    : process.platform === 'win32'
-      ? 'cmd'
-      : 'xdg-open';
-  const args = process.platform === 'win32' ? ['/c', 'start', '', url] : [url];
-  const child = spawn(command, args, { detached: true, stdio: 'ignore' });
-  child.unref();
-}
+export { openBrowser } from '@myco-shared/index.js';
