+++ tsconfig.json
@@ -19,6 +19,7 @@
       "@myco-team-worker/*": ["./packages/myco-team/worker/src/*"],
       "@myco-collective/*": ["./packages/myco-collective/src/*"],
       "@myco-hub/*": ["./packages/myco-hub/src/*"],
+      "@myco-shared/*": ["./packages/myco-shared/src/*"],
       "@myco-deploy/*": ["./packages/myco-deploy/src/*"],
       "react": ["./node_modules/react"],
       "react-dom": ["./node_modules/react-dom"],
@@ -29,7 +30,7 @@
       "react-router": ["./node_modules/react-router/dist/development/index.mjs"]
     }
   },
-  "include": ["packages/myco/src/**/*", "packages/myco-team/src/**/*", "packages/myco-collective/src/**/*", "packages/myco-hub/src/**/*", "packages/myco-deploy/src/**/*"],
+  "include": ["packages/myco/src/**/*", "packages/myco-team/src/**/*", "packages/myco-collective/src/**/*", "packages/myco-hub/src/**/*", "packages/myco-shared/src/**/*", "packages/myco-deploy/src/**/*"],
   "exclude": [
     "node_modules",
     "dist",
